Store the error associated with a reserved mount point.
MountManager currently reserves a virtual mount point for a removable
device with an unrecognized or unsupported filesystem and notifies
Chrome an error. When Chrome crashes, the virtual mount point is
preserved but the error information is not. So after Chrome restarts,
the virtual mount point is incorrectly reported as a real mount point.
This CL modifies MountManager to store the error associated with a
reserved mount path, so that the error can be reported to Chrome after
Chrome recovers from a crash.

BUG=chromium-os:21519
TEST=Tested the following on a Cr48:
1. Log in to a new session and then plug in an unpartitioned USB drive.
2. Verify that the USB drive is mounted as a virtual mount point to a
   folder under External Storage in the file browser.
3. ssh into the Chromebook and kill the chrome process with SIGKILL.
4. After Chrome restarts, verify that the unpartitioned USB drive is
   again mounted as a virtual mount point to a folder under External
   Storage in the file browser.
5. Repeat 1-4 with a FAT-formatted USB drive and verify that the mounted
   folder is preserved after Chrome recopvers from a crash.

Change-Id: I4c924a20dca2cffc83e7820317e47af48c0522b7
Reviewed-on: http://gerrit.chromium.org/gerrit/9882
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>