Use udevadm to determine the name when disabling serio_rawN
Formerly, we relied on a Synaptics created file.  Since we've modified
the Synaptics closed source driver to not rely on information files
in order to close a race condition between udev discovering a device
and running the udev rules and the mount of the stateful partition,
we now need to modify xstart.sh to get the information in a similar
fashion.

BUG=chromium-os:20511
TEST=Use crosh tpcontrol command on Alex, verify no change in behaviour

Change-Id: If9d48add3a426fb0228fc13185cae06ac9d08827
Reviewed-on: https://gerrit.chromium.org/gerrit/9921
Commit-Ready: Terry Lambert <tlambert@chromium.org>
Tested-by: Terry Lambert <tlambert@chromium.org>
Reviewed-by: Terry Lambert <tlambert@chromium.org>