backlight_controller: state should be set correctly
A previous patch to support desktop systems without a backlight made
some changes to BacklightController::SetPowerState().  WriteBrightness()
was being called before the new state was set.  This is incorrect, as
WriteBrightness() uses the new state.  This patch fixes that by setting
the state before calling WriteBrightness().

Also, this patch sets the power state for desktop systems as well.

BUG=chromium-os:21699
TEST=On battery power, set to full brightness, idle to dim, and touch
the trackpad.  The backlight should turn back to the bright level.

Change-Id: Ief2b807b70fca70c5b35d49fae44ca2c1b5ba985
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/10133
Reviewed-by: Sameer Nanda <snanda@chromium.org>