powerd: quiet period required before suspend
Currently, there is a flaw in the way audio detection is done for
suspend timeouts.  If audio is detected to be playing when the system
wants to suspend, it will wait a certain amount of time before
attempting to suspend again.  But if there is a brief pause in the audio
at that time (e.g. changing tracks), it will suspend before the audio
has a chance to pick up again.

This patch fixes that problem by checking for a period of no audio,
defined by react_ms (default 30 sec).  Suspend will happen only if a
period of no audio is detected.  The audio polling will be done at one-
second intervals.  If audio is detected, the suspend will be postponed.

All this is done with a new class, AudioDetector.

Also, the old [Audio/Video]DetectorInterface classes have been replaced
with a single ActivityDetectorInterface class.

BUG=chromium-os:17878
TEST=In /var/lib/power_manager, do:
  echo 5000 > react_ms
  echo 5000 > unplugged_dim_ms
  echo 10000 > unplugged_off_ms
  echo 15000 > unplugged_suspend_ms
  echo 0 > disable_idle_suspend
  restart powerd
Unplug power.  Start playing an audio file for 25 sec:
  aplay -d 25 [audio file]
The system should suspend about 5 seconds after the audio finishes.

Change-Id: I529de2d6e6c55d7809d83d6b6176b28311665576
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/10212