Do not adjust plugged/unplugged brightness offsets the first time
When OnPlugEvent() is first called, it is not a transition from the
other valid plugged/unplugged state.  Instead, it is meant to be an
initialization call.  The plugged state starts at kPowerUnknown.

On a normal transition, OnPlugEvent will adjust the brightness offset of
the new state so that plugged >= unplugged.  This is to avoid the odd
effects of the screen getting brighter when going unplugged and getting
dimmer when getting plugged.  But this shouldn't happen on the initial
call.  Otherwise, suppose you have the following scenario:

Read from prefs:
 plugged = 25
 unplugged = 100
Start powerd on AC, and plugged will be adjusted to 100.  The old
brightness is lost.

The reason unplugged = 100 in the first place is because the adjustment
occurs when transitioning rather than when brightness is changed.  So
it's possible for unplugged to be set to 100 and stay that way while
plugged is set to 25.

To solve this, make sure that the adjustment only takes place when
transitioning from a valid plugged state (i.e. not kPowerUnknown).  This
allows the initialization call to bypass the adjustment.

BUG=chromium-os:21811
TEST=Follow these steps:
1. stop powerd
2. echo 25 > /var/lib/power_manager/plugged_brightness_offset
3. echo 100 > /var/lib/power_manager/unplugged_brightness_offset
4. Plug in AC.
5. restart powerd, while covering the ALS with your thumb.
6. Check that brightness is at about 25% instead of 100%.
7. stop powerd
8. Unplug AC.
9. restart powerd while covering the ALS with your thumb.
10. Check that brightness is at about 100% instead of 25%.

Change-Id: I5422ccaf7d8b4be1b884e7320b18281157524cdb
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/10453
Reviewed-by: Benson Leung <bleung@chromium.org>