cryptohome: bind-mount user and root dirs into place.
This makes them accessible for daemons using the GetDaemonPath() function.

As part of this, add a stack of mountpoints to the UserSession class, which
Mount uses to keep track of which directories are mounted by that UserSession.
This data structure CHECKs that mounts are not leaked.

As a side-effect, make mount.cc cpplint-clean.

BUG=chromium-os:16681
TESTED_ON=x86-generic
TEST=platform_CryptohomeMultiple,unittests,trybot

Try: http://naya.i.corp.google.com/tries/9fcdab92-766e-4325-af46-d77548c5b734.txt
Change-Id: Iecb582d509ec8f792913ba1989769403f12c63ea
Reviewed-On: https://gerrit.chromium.org/gerrit/10670
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/11914
Reviewed-by: Will Drewry <wad@chromium.org>