cryptohome: bind-mount user and root dirs into place.
This makes them accessible for daemons using the GetDaemonPath() function.

As part of this, add a stack of mountpoints to the UserSession class, which
Mount uses to keep track of which directories are mounted by that UserSession.
This data structure CHECKs that mounts are not leaked.

As a side-effect, make mount.cc cpplint-clean.

BUG=chromium-os:16681
TEST=platform_CryptohomeMultiple,unittests

Change-Id: I1ba09af66d91c2d0cc1856d3265a659f6d9b14e8
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/10670
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>