Add IxChariot Performance Endpoint to test images
This CL checks to see if the IxChariot Performance Endpoint is installed.  (It
should be installed in test images only.)  If it is installed, it starts the
endpoint and opens a TCP port (using iptables) for controlling it.

BUG=chromium-os:21991
TEST=Manually started a test image and verified that the endpoint starts and the
TCP port is opened.  Verified on a non-test image that the TCP port is not
opened.

Change-Id: I204761d015e774142653ada88e474926043c14c8
Reviewed-on: https://gerrit.chromium.org/gerrit/10595
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>