[login_manager] s/,/;/ in HandleChildExit
In a couple places in HandleChrildExit, a LOG() ended with a comma,
which seems to be acceptable in a stream operator context and
simply mean that the command on the next line will be run
before the logging happens.  This was leading to confusing log behavior
and is probably just bad to allow to stand in general.

BUG=None
TEST=compile, unit

Change-Id: I88aab828f7007144e447e7df4fc54bc32cedee15
Reviewed-on: http://gerrit.chromium.org/gerrit/10657
Commit-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>