Move the date sanity check so it occurs earlier
This patch moves the date sanity check to occur before htpdate. It
turns out that sometimes "fixing" the date is not enough to make
SysV semaphores work. So in order to avoid ugly hacks to remove old
semaphores, we will simply move the sanity check to an earlier part
of system boot.

Additionally, if the date needs to be reset, it sets it to Jan 2
(instead of Jan 1) 1970 to account for discrepancies between UTC
and whatever timezone the system is set in.

BUG=chromium-os:13200
TEST=Tested on Stumpy w/ modified firmware that corrupts RTC upon boot

Change-Id: I668ada23e4cba647e42852a4b7fca73c6b45820f
Reviewed-on: https://gerrit.chromium.org/gerrit/12041
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>