Adding BatteryRemainingAtStartOfSession* metrics.
Created function for generating the metric. Added in calls to the metric in the
OnSessionStateChange method. Added unit tests for metric generation
function. Also added in a utility method to the until tests to mirror the
specific SendUMA method being used in the generation code. This utility code is
the same as that in CL 12103, so once one of these CLs lands the other will need
to be rebase

BUG=chromium-os:23437
TEST=Ran new unit tests on x86-alex.
     Built image for x86-alex and tested that the metrics are populating.

Change-Id: I7bbd832939394ff8b5db0a779a5d75dcd51ffb3b
Reviewed-on: https://gerrit.chromium.org/gerrit/12109
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>