Correct tests to make sure that the order of calls is as expected.
Since we are running the tests in a for loop and the expectations don't get
evaluated until the end we have no guaratee that expectation and the call line
up. Adding in a verify and clear call to flush things on each iteration. This
should resolve the issue.

BUG=chromium-os:23645
TEST=Built unit tests and confirmed that they still pass as expected.

Change-Id: I63654f06bb63c310040bd86c2a3dbb0806402e79
Reviewed-on: https://gerrit.chromium.org/gerrit/12330
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>