Create metric for measuring the number of times the user adjusted the brightness during a session.
This metric is irregardless of whether it is an increase or a decrease. A
counter variable is used to track the number of adjustments that occur during a
session.  The metric is generated from this count by making a call to the end of
session generation function which in turn calls the metric specific generation
function.

BUG=chromium-os:23630
TEST=Ran new unit tests
     Built and tested that metric occurs correctly on x86-alex

Change-Id: I2c3b60fc58ce08f8a172b4d9e8867116306251b2
Reviewed-on: https://gerrit.chromium.org/gerrit/12422
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>