Create BatteryRemainingWhenChargeStarts power metric
This metric should only be generated when we have just transitioned to being
connected and we have a battery present. The check for this at the beginning of
the generation method and if the condition is not true exit silently. We could
position the state check code before the call to the generation method in the
calling function, but there are two major reasons for not doing that. First,
that would mean that everywhere we called the method we would have to clone that
logic. In addition the calling function isn't under test and is non-trivial to
bring under test. Thus it is simpler and cleaner to do the state checking in the
generation method. This means that unlike other generation methods it does not
return anything, since if there is a fault it will emit the error message.

If we are making a transition from Unknown to Connected we do not generate an
event, since in this case we are likely starting a session, so we don't know how
long the device was actually charging for before hand. This also prevent
spurious events from logging in and out, etc. Thus we have a really simple check
in the calling function to prevent the kPowerUnknown -> kPowerConnected
transition being reported.

BUG=chromium-os:23638
TEST=Ran new unit tests.
     Built image and tested on x86-alex.
     Confirmed that metric was only populating in histograms when I plugged in
     the device and there were no spurious events from logging out then in.

Change-Id: Ice474c48139bc16e7da04862c301d92ba0025f58
Reviewed-on: https://gerrit.chromium.org/gerrit/12579
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>