[login_manager] Clean up after keygen
When generating a new keypair to bless a new owner, keygen
dumps a scratch file in the user's cryptohome.  If we don't
clean it up, we can run into problems trying to re-own the device
for that owner.  Most users will never run into this, but it
is a problem for devs.

BUG=chromium-os:24017
TEST=unit; manually log in, allow ownership to happen.  stop ui, rm ownership files, start ui, log in.  Ownership should happen successfully.

Change-Id: Ic08267d91adad2ea31e181abaa41ba7fe3461faa
Reviewed-on: https://gerrit.chromium.org/gerrit/12793
Commit-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Nikita Kostylev <nkostylev@chromium.org>