Create NumberOfSessionsPerCharge Metric
This CL adds in the NumberOfSessionsPerCharge Metric and utility class to
support it. The utility class is called MetricsStore and provides a persistant
mmap based storage mechanism for having metric counts continue across
reboots. The only metric that currently uses it is introduced in this CL. This
CL adds in build rules for the metrics_store* files and unit tests for the new
code.

BUG=chromium-os:23632
TEST=Ran new unittests
     Built and tested behaviour on x86-alex

Change-Id: Ide59cb0d4ba84009b6ef670d0e9bdb01d6402168
Reviewed-on: https://gerrit.chromium.org/gerrit/12938
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>