Detect and handle the unmounting of an Android USB storage gracefully.
When using the Android user interface to unmount a USB storage of an
Android device that is mounted to a Chromebook, the Chromebook does not
receive a udev remove event on the mounted device (e.g. /dev/sdb1) and
thus cannot tell the file manager to unmount the device automatically.
As a workaround, this CL changes DiskManager to listen to the udev
change event on the parent (e.g. /dev/sdb) of the mounted device and
check if it has any available media to determine if the USB storage has
been unmounted from Android.

This CL makes the following changes:
1. Change DeviceEventSourceInterface to return a list of DeviceEvent
   through its GetDeviceEvents() method, which will be dispatched in
   sequence by DeviceEventModerator::ProcessDeviceEvents().
2. Remove the DeviceEventAddedAfterRemoved event type as
   DeviceEventSourceInterface can now return multiple events per
   invocation of GetDeviceEvents().
3. Change DiskManager, upon receiving a media change event on a block
   device, to dispatch a kDiskRemoved event for each immediate child
   of that block device if it has no available media.
4. Modify unit tests accordingly.

BUG=chromium-os:23858
TEST=Tested the following:
1. Build cros-disks for x86-generic, amd64-generic, and arm-generic.
2. Run cros-disks unit tests.
3. Manually test mounting, unmounting, and remounting a USB drive,
   SD card, removable CD-ROM drive, and Android USB storage on a Cr48.

Change-Id: I6bce3fbc3044f3cba2dead2bf468d7d6b6d6b5dc
Reviewed-on: https://gerrit.chromium.org/gerrit/12964
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>