shill: Ignore WiFi devices in "monitor" mode
Ignore WiFi devices setup for capture.  These are useful
for debugging, and shill should not try to manage them as
client interfaces or cause wpa_supplicant to try to convert
them.  In the future we could use these internall in shill
for debugging, so create a new interface type for them.

BUG=chromium-os:24289
TEST=Start up with mon0, install mon0 while shill is running

Change-Id: I91e2d2c4a691d2cf9b25588a6ce6675e60b73df9
Reviewed-on: https://gerrit.chromium.org/gerrit/12991
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>