Prevent under/over flow in LengthOfSession metric
This CL adds in bounds checking to the generation of the LengthOfSession
metricSpecifically it does not send any values thate are negative, since they
are invalid. For values over the limit of 12 hours it clamps them to be 12
hours, which effectively makes the final bucket a "12 hours or more" bucket.

BUG=chromium-os:24818
TEST=Added unit tests to confirm behaviour

Change-Id: I87dd34a552db711df381ece146eb26cc7f18343a
Reviewed-on: https://gerrit.chromium.org/gerrit/13825
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>