power: Add min_backlight_level setting.
This replaces min_backlight_percent, which doesn't look like
it's been used in the nine months since it was added.

Being able to specify a hardware-specific level rather than
a percent seems more precise if we need to use this for
specific hardware, and it fits better with future changes
that I plan to make to adjust the default minimum visible
level and use a non-linear mapping between percent and
hardware levels.

BUG=chromium-os:12103,chromium-os:24569
TEST=added a new test; existing tests pass; also did manual testing after setting a higher minimum level

Change-Id: I7fc6f4813706159e5111ecfa128478ea4331d038
Reviewed-on: https://gerrit.chromium.org/gerrit/13938
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Ready: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>