[cromo] use a local DBus.Properties, not the system one.
The system one is about to become the canonical org.freedesktop.DBus.Properties
interface. The ModemManager API depends on a fake version of that interface
which predates the standardization of the canonical one, so drop XML for the
fake one into the cromo source tree and use that instead of the system one.

BUILD=chromium-os:23839
TEST=build
emerge-$board cromo gobi-cromo-plugin

Change-Id: Iefeb401ecb95fcc277cbf785084176ab3a91121e
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/13982
Reviewed-by: Nathan J. Williams <njw@chromium.org>