vpn-manager: Make rekey default to yes in ipsec.conf
This CL changes the rekey option in IpsecManager default to true such
that the pluto daemon on the client side (Chromium OS) can request a
renegotiation of the connection when it is about to expire. According to
the ipsec.conf man page, setting rekey=no is largely ineffective unless
both ends of the connection agree on it.

BUG=chromium-os:25070
TEST=Verified the following:
1. Connect successfully to Windows 2008 RRAS server, Cisco ASA 5505
   and StrongSWAN VPN server with a L2TP/IPsec pre-shared key or user
   certificate.
2. Check that pluto requests a renegotiation of the connection when the
   connection is about to expire, and the connection remains established
   after that.
3. Run network_VPN autotest test suite.

Change-Id: Ic9c25df0ff3fd3329e0ec3bccea4eade51a1aa54
Reviewed-on: https://gerrit.chromium.org/gerrit/14197
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Ben Chan <benchan@chromium.org>