backlight_controller: fix strange ALS-idle behavior
The problem:
1. ALS is covered, and the brightness falls to the min. It would
fall to 0 if it were not for the existing check.
2. powerd goes to idle-dim. No brightness change here, as the
backlight is already dimmed at the min level.
3. User input brings powerd out of idle-dim state, into active
state. With the ALS still at a low level, the target percent is
calculated to be 0%.
4. As a result, the brightness drops to 0 when the system comes out
of idle.  This is a bad effect.

BUG=chromium-os:25102
TEST=1. Set backlight rather low, like 15%.
2. Cover the light sensor or enter a dark room.  Backlight should drop
below 10%.  The initial backlight setting in step 1 may need to be
adjusted to achieve this.
3. Wait for the system to Idle.  No change is observed since we are
already under 10%, the backlight idle level.  I set my
/usr/shared/power_manager/unplugged_dim_ms to 30000 to Idle in 30
seconds for convenience.
4. Uncover light sensor, or return to normal light.
5. Swipe finger on touchpad or touch keyboard to wake
6. The backlight should not turn off.  After a few seconds, it should
increase.

Change-Id: I6782620123ea1892033e857135d34476efdf2474
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/14649