Add protobuf based code path for generation of power supply properties
This CL adds in SCons rules to pull in the proto defintion from system_api and
generate the source files for the protobuf. The protobuf is used in the response
to the GetPowerSupplyProperties signal via dbus to create an alternative code path to
send up information. It is intended that both code paths will briefly exist and
then the old one will be deprecated.

BUG=chromium-os:25489
TEST=This is a D-Bus change and I am not really sure if I can/how to unit test
     Tested with Chrome side changes on custom build to make sure there are no
     performance or behaviour regressions.

Change-Id: Ic3abc9c9606c8ca0ff96ab96747070d6b23f2626
Reviewed-on: https://gerrit.chromium.org/gerrit/14901
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>