backlight_controller: restore brightness on resume
The backlight controller currently does not handle resumes seamlessly.
It leaves it up to the backlight driver to set the backlight level.

This patch introduces a way to save the value of the backlight before
leaving the active level.  Later, when backlight controller returns to
the active state from suspend, the saved offset value is restored.

Further changes:
1. Provides a flag for the WriteBrightness function to choose gradual vs
instant backlight adjustments.
2. On resume, setting the backlight state to active is done as the last
step, to avoid collisions with other system calls that will update the
state of the backlight (e.g. MonitorReconfigure::Run()).
3. Suspend from lid sets the backlight state to suspended, so it knows
to save the active level.

BUG=chrome-os-partner:7680
TEST=Set backlight to something between 0 and 100%, like 25% or 50%.
Close the lid to suspend.  Open it and resume.  The backlight should be
restored.
Set the following in /var/lib/power_manager:
echo 1000 > react_ms
echo 0 > disable_idle_suspend
echo 5000 > unplugged_dim_ms
echo 10000 > unplugged_off_ms
echo 15000 > unplugged_suspend_ms
Let it go to idle-suspend, and resume.  The old backlight level (in the
middle of the scale) should be restored.

Change-Id: I2360b2af5855338c2017942d349e089808d22255
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/14858