[cryptohome] refactor service
Refactor the service class to not reach into the mount class and use its tpm
when it needs a reference to the tpm. This is the first stage in the refactoring
needed to have cryptohome handle mounts for multiple users at once, since
instead of needing to have a mount present to touch the tpm, we have our own
reference to it.

BUG=chromium-os:20778,chromium-os:20183
TEST=unit

Change-Id: I334b98b8e2cb338b87e16b99ea1b63cc60e40c05
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/15201