shill: Fix unit test by removing "Uninteresting" calls to mock methods.
When running the wifi_unittest, there are several tests that pass but they emit
messages about uninteresting calls to mock methods; that is, calls to mock
methods for which no expectation has been established.  These messages are noisy
and may mask a problem in a test where a mock method should or shouild not have
been called.  It is better to not have any messages about uninteresting calls to
mock methods.

BUG=none
TEST=All shill unit tests pass.

Change-Id: Ia2ff0995d26b8db3d25ff08e558c17243ab1e009
Reviewed-on: https://gerrit.chromium.org/gerrit/15344
Tested-by: Gary Morain <gmorain@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Gary Morain <gmorain@chromium.org>